requirements.txt
pyautogui==0.9.54
pyserial==3.5
